requirements.txt
llama-cpp-python
whisper-cpp-python
fastapi[standard]
sse-starlette
uvicorn